rename test_cases -> common_tests_cases
