+++ ydb/library/yql/providers/generic/connector/tests/common_test_cases/select_missing_table.py
@@ -1,7 +1,7 @@
 from typing import List
 
 from ydb.library.yql.providers.generic.connector.api.common.data_source_pb2 import EDataSourceKind, EProtocol
-from ydb.library.yql.providers.generic.connector.tests.test_cases.base import BaseTestCase
+from ydb.library.yql.providers.generic.connector.tests.common_test_cases.base import BaseTestCase
 
 
 TestCase = BaseTestCase
@@ -11,7 +11,7 @@ class Factory:
     def make_test_cases(self, data_source_kind: EDataSourceKind) -> List[TestCase]:
         test_cases = []
 
-        test_case_name = 'missing_database'
+        test_case_name = 'missing_table'
 
         test_case = TestCase(
             name_=test_case_name,
