+++ ydb/library/yql/providers/generic/connector/tests/common_test_cases/select_positive_common.py
@@ -18,7 +18,7 @@ from ydb.library.yql.providers.generic.connector.tests.utils.schema import (
     SelectWhere,
 )
 
-from ydb.library.yql.providers.generic.connector.tests.test_cases.base import BaseTestCase
+from ydb.library.yql.providers.generic.connector.tests.common_test_cases.base import BaseTestCase
 from ydb.library.yql.providers.generic.connector.tests.utils.settings import GenericSettings
