+++ ydb/library/yql/providers/generic/connector/tests/datasource/clickhouse/ya.make
@@ -50,7 +50,7 @@ PEERDIR(
     contrib/python/pytest
     ydb/library/yql/providers/generic/connector/api/common
     ydb/library/yql/providers/generic/connector/api/service/protos
-    ydb/library/yql/providers/generic/connector/tests/test_cases
+    ydb/library/yql/providers/generic/connector/tests/common_test_cases
     ydb/library/yql/providers/generic/connector/tests/utils
     ydb/library/yql/providers/generic/connector/tests/utils/clients
     ydb/library/yql/providers/generic/connector/tests/utils/scenario
