+++ ydb/library/yql/providers/generic/connector/tests/datasource/postgresql/ya.make
@@ -50,7 +50,7 @@ TEST_SRCS(
 PEERDIR(
     contrib/python/pytest
     ydb/library/yql/providers/generic/connector/api/common
-    ydb/library/yql/providers/generic/connector/tests/test_cases
+    ydb/library/yql/providers/generic/connector/tests/common_test_cases
     ydb/library/yql/providers/generic/connector/tests/utils
     ydb/library/yql/providers/generic/connector/tests/utils/types
     ydb/library/yql/providers/generic/connector/tests/utils/clients
