+++ ydb/library/yql/providers/generic/connector/tests/ya.make
@@ -1,6 +1,6 @@
 RECURSE_FOR_TESTS(
+    common_test_cases
     datasource
     join
-    test_cases
     utils
 )
